Reorganized custom hooks storage
