+++ src/components/layouts/appTable/decksTableBody.tsx
@@ -3,7 +3,7 @@ import { Link } from 'react-router-dom'
 import EditIcon from '@/assets/svg/editIcon.svg?react'
 import PlayIcon from '@/assets/svg/playIcon.svg?react'
 import { formatDate } from '@/common/commonFunctions'
-import { useModalStateHandler } from '@/common/customHooks'
+import { useModalStateHandler } from '@/common/customHooks/useModalStateHandler'
 import { ConfirmDeleteModal } from '@/components/layouts/modals/confirmDeleteModal/confirmDeleteModal'
 import { EditDeckModal } from '@/components/layouts/modals/editDeck/editDeck'
 import { Deck, UpdateDeckArgs } from '@/services/decks/decks.types'
